change: move GetVersion() from mp4 to internal
