+++ cmd/mp4ff-crop/main.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"sort"
 
+	"github.com/Eyevinn/mp4ff/internal"
 	"github.com/Eyevinn/mp4ff/mp4"
 )
 
@@ -62,7 +63,7 @@ func run(args []string, stdout io.Writer) error {
 	}
 
 	if o.version {
-		fmt.Fprintf(stdout, "%s %s\n", appName, mp4.GetVersion())
+		fmt.Fprintf(stdout, "%s %s\n", appName, internal.GetVersion())
 		return nil
 	}
