+++ cmd/mp4ff-info/main.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/Eyevinn/mp4ff/internal"
 	"github.com/Eyevinn/mp4ff/mp4"
 )
 
@@ -59,7 +60,7 @@ func run(args []string, w io.Writer) error {
 	}
 
 	if opts.version {
-		fmt.Printf("%s %s\n", appName, mp4.GetVersion())
+		fmt.Printf("%s %s\n", appName, internal.GetVersion())
 		return nil
 	}
