+++ cmd/mp4ff-subslister/main.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/Eyevinn/mp4ff/internal"
 	"github.com/Eyevinn/mp4ff/mp4"
 )
 
@@ -64,7 +65,7 @@ func run(args []string, stdout io.Writer) error {
 	}
 
 	if o.version {
-		fmt.Fprintf(stdout, "%s %s\n", appName, mp4.GetVersion())
+		fmt.Fprintf(stdout, "%s %s\n", appName, internal.GetVersion())
 		return nil
 	}
