+++ internal/version.go
@@ -1,4 +1,4 @@
-package mp4
+package internal
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 
 var (
 	commitVersion string = "v0.46"      // May be updated using build flags
-	commitDate    string = "1723143113" // commitDate in Epoch seconds (may be overridden using build flags)
+	commitDate    string = "1731409630" // commitDate in Epoch seconds (may be overridden using build flags)
 )
 
 // GetVersion - get version and also commitHash and commitDate if inserted via Makefile
