+++ internal/version_test.go
@@ -1,8 +1,8 @@
-package mp4
+package internal
 
 import "fmt"
 
 func ExampleGetVersion() {
 	fmt.Println(GetVersion())
-	// Output: v0.46, date: 2024-08-08
+	// Output: v0.46, date: 2024-11-12
 }
